Refactor sprite handling and input functions
- Updated sprite loading mechanism to utilize SRL bitmap loader and VDP1 texture management.
- Changed sprite array from pointer to static array for better memory management.
- Enhanced logging for sprite loading process to include debug and trace information.
- Modified drawSprite function signature to use uint8_t for sprite index and int16_t for position parameters.
- Cleaned up commented-out code and improved overall readability.
- Adjusted palette initialization and smoke buffer creation for consistency.

+++ src/screen.h
@@ -65,7 +65,7 @@ void drawBoxPanel(int x, int y, int width, int height,
 int drawNum(int n, int x ,int y, int s, int c1, int c2);
 int drawNumRight(int n, int x ,int y, int s, int c1, int c2);
 int drawNumCenter(int n, int x ,int y, int s, int c1, int c2);
-void drawSprite(const int n, const int x, const int y);
+void drawSprite(const uint8_t n, const int16_t x, const int16_t y);
 
 int getPadState();
 int getButtonState();
